+++ firebase-functions/functions/src/models/player.ts
@@ -1,0 +1,4 @@
+﻿export interface Player {
+  choice: 'Rock' | 'Paper' | 'Scissors',
+  active: boolean;
+}
